+++ packages/@types/src/models/Space.ts
@@ -1,3 +1,5 @@
+import Stripe from "stripe";
+
 import Actions from "@sentrei/types/models/Actions";
 import Analytics from "@sentrei/types/models/Analytics";
 import Metadata from "@sentrei/types/models/Metadata";
@@ -20,6 +22,9 @@ declare namespace Space {
     stats: Stats.Fields;
     tier: Tiers;
     stripeId: string;
+    subscriptionId?: string;
+    subscriptionStatus?: Stripe.Subscription.Status;
+    subscriptionTrial?: boolean;
   }
 
   export type AdminUpdate = Partial<Fields>;